requirements.txt
accelerate==1.7.0
altair==5.5.0
attrs==25.3.0
bitsandbytes==0.45.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dateparser==1.2.1
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==1.40.0
networkx==3.3
numpy==2.1.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.0
safetensors==0.5.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
streamlit-calendar==1.3.1
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tornado==6.5
tqdm==4.67.1
transformers==4.52.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
watchdog==6.0.0
